+++ scripts/TabSigne.py
@@ -31,4 +31,4 @@ A = parser.parse_args()
 tmp = TableauSigne(sympify(A.expr), bornes = eval(A.bornes))
 # choix du format dans le dictionnaire fin des 2 méthodes.
 fin = {'tex':tmp.export_latex, 'pst':tmp.export_pst, 'pag':tmp.export_pst}
-fin[A.format](nom = A.out, simplif = False, ext = A.format) 
+fin[A.format](nom = A.out, option='whole', ext = A.format) #simplif = False